+++ intake.txt
@@ -1,4 +1,3 @@
-<<<<<<< HEAD
 GBF2402EN;Foundation,Engineering;12;L002;L004;2023-12-01;2024-02-25;2024-03-01;2024-06-30
 GBF2402IS;Foundation;Intelligent System;12;L014;L007;2023-12-01;2024-02-25;2024-03-01;2024-06-30
 GBF2402BK;Foundation;Banking;12;L007;L014;2023-12-01;2024-02-25;2024-03-01;2024-06-30
@@ -18,24 +17,3 @@ GBD2407CS;Degree;Cyber Security;36;L010;L005;2024-04-01;2024-07-05;2024-07-12;20
 GBM2402FT;Master;FinTech;12;L015;L002;2023-12-01;2024-02-25;2024-03-01;2024-06-30
 GBM2402IS;Master;Intelligent System;12;L014;L008;2023-12-01;2024-02-25;2024-03-01;2024-06-30
 GBM2407PS;Master;Psychology;12;L008;L013;2024-04-01;2024-07-05;2024-07-12;2024-11-25
-=======
-GBF2402EN;Foundation;Engineering;12;L002;L007;2023-12-01;2024-02-25;2024-03-01;2024-06-30
-GBF2402IS;Foundation;Intelligent System;12;L014;L010;2023-12-01;2024-02-25;2024-03-01;2024-06-30
-GBF2402BK;Foundation;Banking;12;L007;L014;2023-12-01;2024-02-25;2024-03-01;2024-06-30
-GBF2407FT;Foundation;FinTech;12;L003;L015;2024-04-01;2024-07-05;2024-07-12;2024-11-25
-GBF2407BN;Foundation;Business;12;L001;L005;2024-04-01;2024-07-05;2024-07-12;2024-11-25
-GBU2302AC;Diploma;Accounting;24;L006;L009;2022-12-01;2023-02-25;2023-03-01;2023-06-30
-GBU2402EN;Diploma;Engineering;24;L007;L002;2023-12-01;2024-02-25;2024-03-01;2024-06-30
-GBU2402CE;Diploma;Cloud Engineering;24;L009;L001;2023-12-01;2024-02-25;2024-03-01;2024-06-30
-GBU2407FN;Diploma;Finance;24;L008;L013;2024-04-01;2024-07-05;2024-07-12;2024-11-25
-GBU2407DM;Diploma;Digital Marketing;24;L013;L003;2024-04-01;2024-07-05;2024-07-12;2024-11-25
-GBU2407DS;Diploma;Design;24;L004;L012;2024-04-01;2024-07-05;2024-07-12;2024-11-25
-GBD2302FN;Degree;Finance;24;L013;L008;2022-12-01;2023-02-25;2023-03-01;2023-06-30
-GBD2402SE;Degree;Software Engineering;36;L015;L011;2023-12-01;2024-02-25;2024-03-01;2024-06-30
-GBD2407DA;Degree;Data Analytics;36;L011;L006;2024-04-01;2024-07-05;2024-07-12;2024-11-25
-GBD2407DF;Degree;Digital Forensics;36;L005;L012;2024-04-01;2024-07-05;2024-07-12;2024-11-25
-GBD2407CS;Degree;Cyber Security;36;L010;L004;2024-04-01;2024-07-05;2024-07-12;2024-11-25
-GBM2402FT;Master;FinTech;12;L015;L003;2023-12-01;2024-02-25;2024-03-01;2024-06-30
-GBM2402IS;Master;Intelligent System;12;L014;L010;2023-12-01;2024-02-25;2024-03-01;2024-06-30
-GBM2407PS;Master;Psychology;12;L008;L002;2024-04-01;2024-07-05;2024-07-12;2024-11-25
->>>>>>> main
