+++ nbproject/project.properties
@@ -35,7 +35,6 @@ dist.jlink.dir=${dist.dir}/jlink
 dist.jlink.output=${dist.jlink.dir}/GoodBrain
 endorsed.classpath=
 excludes=
-<<<<<<< HEAD
 file.reference.flatlaf-3.2.5.jar=C:\\Users\\User\\Downloads\\flatlaf-3.2.5.jar
 file.reference.flatlaf-3.2.5.jar-1=C:\\Users\\User\\Downloads\\Java library\\flatlaf-3.2.5.jar
 file.reference.flatlaf-3.4.1.jar=C:\\Users\\User\\Downloads\\flatlaf-3.4.1.jar
@@ -79,6 +78,7 @@ javac.classpath=\
     ${libs.JCalendar.classpath}:\
     ${libs.absolutelayout.classpath}:\
     ${file.reference.itextpdf-5.5.13.3.jar}
+<<<<<<< HEAD
 =======
 file.reference.jcalendar-1.4.jar=C:\\Users\\User\\Downloads\\lib\\jcalendar-1.4.jar
 file.reference.TimingFramework-0.55.jar=C:\\Users\\User\\Downloads\\TimingFramework-0.55.jar
@@ -101,6 +101,8 @@ javac.classpath=\
     ${file.reference.swing-toast-notifications-1.0.1.jar}:\
     ${file.reference.jcalendar-1.4.jar-1}
 >>>>>>> main
+=======
+>>>>>>> parent of 3af4316 (merge main into admin)
 # Space-separated list of extra javac options
 javac.compilerargs=
 javac.deprecation=false
