+++ presentation.txt
@@ -1,7 +1,7 @@
-PR0001;S001;FYPGBU2302AC;2023-06-13 09:30;Pending;L006;Accept;L009;Accept
-PR0002;S003;FYPGBU2302AC;2023-06-13 10:30;Accept;L006;Accept;L009;Accept
-PR0003;S004;FYPGBU2302AC;2023-06-13 11:00;Accept;L006;Accept;L009;Accept
-PR0004;S007;INTGBD2302FN;2024-03-26 18:22;Accept;L013;Accept;L008;Accept
-PR0005;S005;INTGBD2302FN;2024-03-26 08:30;Accept;L013;Accept;L008;Accept
-PR0006;S009;INTGBD2302FN;2024-03-26 09:00;Accept;L013;Accept;L008;Accept
-PR0007;S014;INTGBD2302FN;2024-03-26 10:30;Pending;L013;Pending;L008;Accept
+PR0001,S001,FYPGBU2302AC,2023-06-13 09:30,L006,Changed,2023-06-23 09:30,L009,Yes,2023-06-13 09:30
+PR0002,S003,FYPGBU2302AC,2023-06-13 10:30,Accept,L006,Accept,L009,Accept
+PR0003,S004,FYPGBU2302AC,2023-06-13 11:00,Accept,L006,Accept,L009,Accept
+PR0004,S007,INTGBD2302FN,2024-03-26 18:22,Accept,L003,Accept,L010,Accept
+PR0005,S005,INTGBD2302FN,2024-03-26 08:30,Accept,L003,Accept,L010,Accept
+PR0006,S009,INTGBD2302FN,2024-03-26 09:00,Accept,L003,Accept,L010,Accept
+PR0007,S014,INTGBD2302FN,2024-03-26 10:30,Pending,L003,Pending,L010,Accept
