+++ project.txt
@@ -1,6 +1,8 @@
-FYPGBU2302AC;Final Year Project;GBU2302AC;L006;L009;2023-06-15
-RMCPGBM2402FT;RMCP;GBM2402FT;L015;L003;2025-01-15
-CP1GBU2402CE;Capstone Project 1;GBU2402CE;L009;L001;2025-01-31
-CP2GBU2302AC;Capstone Project 2;GBU2302AC;L006;L009;2026-01-31
-IRGBM2402IS;Investigation Report;GBM2402IS;L014;L010;2025-01-31
-INTGBU2407DS;Internships;GBU2407DS;L012;L004;2024-11-25
+FYPGBU2302AC,Final Year Project,GBU2302AC,2023-06-15
+RMCPGBM2402FT,RMCP,GBM2402FT,2025-01-15
+CP1GBU2402CE,Capstone Project 1,GBU2402CE,2025-01-31
+CP2GBU2302AC,Capstone Project 2,GBU2302AC,2026-01-31
+IRGBM2402IS,Investigation Report,GBM2402IS,2025-01-31
+INTGBD2302FN,Internship,GBD2302FN,2024-05-31
+
+
