+++ src/MainProgram/loginPage.java
@@ -6,7 +6,6 @@ package MainProgram;
 
 import javax.swing.Icon;
 import javax.swing.JOptionPane;
-import project_manager.pmHomePage;
 
 public class loginPage extends javax.swing.JFrame {
 
@@ -114,7 +113,7 @@ public class loginPage extends javax.swing.JFrame {
         int role = check.checkCredentials(id, password);
         
         if (role != -1){
-            showPage(role, id);
+            showPage(role);
         }
         else{
              JOptionPane.showMessageDialog(null, "Invalid user ID or password.", "Error",JOptionPane.ERROR_MESSAGE);
@@ -123,7 +122,7 @@ public class loginPage extends javax.swing.JFrame {
         }
     }//GEN-LAST:event_jButton1ActionPerformed
 
-    private void showPage(int role, String id){
+    private void showPage(int role){
         Icon icon = new javax.swing.ImageIcon(getClass().getResource("/Icon/success.png"));
         JOptionPane.showMessageDialog(null, "Log In Success!", "Notification", JOptionPane.INFORMATION_MESSAGE, icon);
         
@@ -145,7 +144,7 @@ public class loginPage extends javax.swing.JFrame {
                 this.dispose();
                 break;
             case 4:
-                pmHomePage pm = new pmHomePage(id);
+                projectManagerPage pm = new projectManagerPage();
                 pm.setVisible(true);
                 this.dispose();
                 break;
