--- src/actionCell/tableActionEvent.java
@@ -1,17 +1,0 @@
-/*
- * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
- * Click nbfs://nbhost/SystemFileSystem/Templates/Classes/Class.java to edit this template
- */
-package actionCell;
-
-/**
- *
- * @author User
- */
-public interface tableActionEvent {
-    
-    public void onEdit(int row);
-    
-    public void onDelete(int row);
-    
-}
