--- src/pmToggle/ToggleAdapter.java
@@ -1,16 +1,0 @@
-/*
- * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
- * Click nbfs://nbhost/SystemFileSystem/Templates/Classes/Class.java to edit this template
- */
-package toggle;
-
-public abstract class ToggleAdapter implements ToggleListener {
-
-    @Override
-    public void onSelected(boolean selected) {
-    }
-
-    @Override
-    public void onAnimated(float animated) {
-    }
-}
