--- src/pmToggle/ToggleListener.java
@@ -1,12 +1,0 @@
-/*
- * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
- * Click nbfs://nbhost/SystemFileSystem/Templates/Classes/Class.java to edit this template
- */
-package toggle;
-
-public interface ToggleListener {
-
-    public void onSelected(boolean selected);
-
-    public void onAnimated(float animated);
-}
