+++ src/project_manager/all_info.java
@@ -4,20 +4,15 @@
  */
 package project_manager;
 
-import java.util.Date;
-
 /**
  *
  * @author User
  */
-public class pmProjectInfo {
+public class all_info {
     private String projectID;
     private String projectName;
     private String intake;
-    private String supervisor;
-    private String secondMarker;
     private String dueDate;
-    
 
     /**
      * @return the projectID
@@ -60,34 +55,6 @@ public class pmProjectInfo {
     public void setIntake(String intake) {
         this.intake = intake;
     }
-    
-        /**
-     * @return the intake
-     */
-    public String getSupervisor() {
-        return supervisor;
-    }
-
-    /**
-     * @param supervisor the intake to set
-     */
-    public void setSupervisor(String supervisor) {
-        this.supervisor = supervisor;
-    }
-    
-        /**
-     * @return the intake
-     */
-    public String getSecondMarker() {
-        return secondMarker;
-    }
-
-    /**
-     * @param secondMarker the intake to set
-     */
-    public void setSecondMarker(String secondMarker) {
-        this.secondMarker = secondMarker;
-    }
 
     /**
      * @return the dueDate
@@ -97,23 +64,10 @@ public class pmProjectInfo {
     }
 
     /**
-     * @param projDueDate the dueDate to set
+     * @param dueDate the dueDate to set
      */
     public void setDueDate(String dueDate) {
         this.dueDate = dueDate;
     }
-    
-    
-//    editpage page = new editpage(info);
-//    
-//    
-//    public editpage(all_info info);info info = new all_info();
-//    info.setDueDate(arraylist[3].toString());
-//    ..
-//    
-//    editpage page = new editpage(info);
-//    
-//    
-//    public editpage(all_info info);
     
 }
