--- src/project_manager/pmIntakeInfo.java
@@ -1,147 +1,0 @@
-/*
- * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
- * Click nbfs://nbhost/SystemFileSystem/Templates/Classes/Class.java to edit this template
- */
-package project_manager;
-
-/**
- *
- * @author User
- */
-public class pmIntakeInfo {
-    private String intakeID;
-    private String intakeName;
-    private String intakeDuration;
-    private String intakeSupervisor;
-    private String intakeSecondMarker;
-    private String intakeRegisterStart;
-    private String intakeRegisterEnd;
-    private String intakeStart;
-    private String intakeEnd;
-
-    /**
-     * @return the intakeID
-     */
-    public String getIntakeID() {
-        return intakeID;
-    }
-
-    /**
-     * @param intakeID the intakeID to set
-     */
-    public void setIntakeID(String intakeID) {
-        this.intakeID = intakeID;
-    }
-
-    /**
-     * @return the intakeName
-     */
-    public String getIntakeName() {
-        return intakeName;
-    }
-
-    /**
-     * @param intakeName the intakeName to set
-     */
-    public void setIntakeName(String intakeName) {
-        this.intakeName = intakeName;
-    }
-
-    /**
-     * @return the intakeDuration
-     */
-    public String getIntakeDuration() {
-        return intakeDuration;
-    }
-
-    /**
-     * @param intakeDuration the intakeDuration to set
-     */
-    public void setIntakeDuration(String intakeDuration) {
-        this.intakeDuration = intakeDuration;
-    }
-
-    /**
-     * @return the intakeSupervisor
-     */
-    public String getIntakeSupervisor() {
-        return intakeSupervisor;
-    }
-
-    /**
-     * @param intakeSupervisor the intakeSupervisor to set
-     */
-    public void setIntakeSupervisor(String intakeSupervisor) {
-        this.intakeSupervisor = intakeSupervisor;
-    }
-
-    /**
-     * @return the intakeSecondMarker
-     */
-    public String getIntakeSecondMarker() {
-        return intakeSecondMarker;
-    }
-
-    /**
-     * @param intakeSecondMarker the intakeSecondMarker to set
-     */
-    public void setIntakeSecondMarker(String intakeSecondMarker) {
-        this.intakeSecondMarker = intakeSecondMarker;
-    }
-
-    /**
-     * @return the intakeRegisterStart
-     */
-    public String getIntakeRegisterStart() {
-        return intakeRegisterStart;
-    }
-
-    /**
-     * @param intakeRegisterStart the intakeRegisterStart to set
-     */
-    public void setIntakeRegisterStart(String intakeRegisterStart) {
-        this.intakeRegisterStart = intakeRegisterStart;
-    }
-
-    /**
-     * @return the intakeRegisterEnd
-     */
-    public String getIntakeRegisterEnd() {
-        return intakeRegisterEnd;
-    }
-
-    /**
-     * @param intakeRegisterEnd the intakeRegisterEnd to set
-     */
-    public void setIntakeRegisterEnd(String intakeRegisterEnd) {
-        this.intakeRegisterEnd = intakeRegisterEnd;
-    }
-
-    /**
-     * @return the intakeStart
-     */
-    public String getIntakeStart() {
-        return intakeStart;
-    }
-
-    /**
-     * @param intakeStart the intakeStart to set
-     */
-    public void setIntakeStart(String intakeStart) {
-        this.intakeStart = intakeStart;
-    }
-
-    /**
-     * @return the intakeEnd
-     */
-    public String getIntakeEnd() {
-        return intakeEnd;
-    }
-
-    /**
-     * @param intakeEnd the intakeEnd to set
-     */
-    public void setIntakeEnd(String intakeEnd) {
-        this.intakeEnd = intakeEnd;
-    }
-}
