--- src/project_manager/pmPresentationInfo.java
@@ -1,148 +1,0 @@
-/*
- * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
- * Click nbfs://nbhost/SystemFileSystem/Templates/Classes/Class.java to edit this template
- */
-package project_manager;
-
-/**
- *
- * @author User
- */
-public class pmPresentationInfo {
-    private String presentationID;
-    private String studentID;
-    private String projectID;
-    private String presentationDateTime;
-    private String studentAcc;
-    private String supervisorID;
-    private String suprvisorAcc;
-    private String secondMarkerID;
-    private String secondMarkerAcc;
-
-    /**
-     * @return the presentationID
-     */
-    public String getPresentationID() {
-        return presentationID;
-    }
-
-    /**
-     * @param presentationID the presentationID to set
-     */
-    public void setPresentationID(String presentationID) {
-        this.presentationID = presentationID;
-    }
-
-    /**
-     * @return the studentID
-     */
-    public String getStudentID() {
-        return studentID;
-    }
-
-    /**
-     * @param studentID the studentID to set
-     */
-    public void setStudentID(String studentID) {
-        this.studentID = studentID;
-    }
-
-    /**
-     * @return the projectID
-     */
-    public String getProjectID() {
-        return projectID;
-    }
-
-    /**
-     * @param projectID the projectID to set
-     */
-    public void setProjectID(String projectID) {
-        this.projectID = projectID;
-    }
-
-    /**
-     * @return the presentationDateTime
-     */
-    public String getPresentationDateTime() {
-        return presentationDateTime;
-    }
-
-    /**
-     * @param presentationDateTime the presentationDateTime to set
-     */
-    public void setPresentationDateTime(String presentationDateTime) {
-        this.presentationDateTime = presentationDateTime;
-    }
-
-    /**
-     * @return the studentAcc
-     */
-    public String getStudentAcc() {
-        return studentAcc;
-    }
-
-    /**
-     * @param studentAcc the studentAcc to set
-     */
-    public void setStudentAcc(String studentAcc) {
-        this.studentAcc = studentAcc;
-    }
-
-    /**
-     * @return the supervisorID
-     */
-    public String getSupervisorID() {
-        return supervisorID;
-    }
-
-    /**
-     * @param supervisorID the supervisorID to set
-     */
-    public void setSupervisorID(String supervisorID) {
-        this.supervisorID = supervisorID;
-    }
-
-    /**
-     * @return the suprvisorAcc
-     */
-    public String getSuprvisorAcc() {
-        return suprvisorAcc;
-    }
-
-    /**
-     * @param suprvisorAcc the suprvisorAcc to set
-     */
-    public void setSuprvisorAcc(String suprvisorAcc) {
-        this.suprvisorAcc = suprvisorAcc;
-    }
-
-    /**
-     * @return the secondMarkerID
-     */
-    public String getSecondMarkerID() {
-        return secondMarkerID;
-    }
-
-    /**
-     * @param secondMarkerID the secondMarkerID to set
-     */
-    public void setSecondMarkerID(String secondMarkerID) {
-        this.secondMarkerID = secondMarkerID;
-    }
-
-    /**
-     * @return the secondMarkerAcc
-     */
-    public String getSecondMarkerAcc() {
-        return secondMarkerAcc;
-    }
-
-    /**
-     * @param secondMarkerAcc the secondMarkerAcc to set
-     */
-    public void setSecondMarkerAcc(String secondMarkerAcc) {
-        this.secondMarkerAcc = secondMarkerAcc;
-    }
-    
-}
