+++ student.txt
@@ -17,6 +17,7 @@ S016;Jaydon Aguirre;040209097571;60134776859;S016@mail.goodbrain.edu.my;GBD2402S
 S017;Lucie Carver;010624546262;60174922416;S017@mail.goodbrain.edu.my;GBD2402SE
 S018;Walter Lawson;010204568838;60102457836;S018@mail.goodbrain.edu.my;GBU2407DM
 S019;Georgiana Dixon;050515274762;60133844963;S019@mail.goodbrain.edu.my;GBD2407DF
+<<<<<<< HEAD
 <<<<<<< Updated upstream
 <<<<<<< HEAD
 S020;Nicholas Patel;020903508054;60138591060;S020@mail.goodbrain.edu.my;GBF2407BNS021;Harry Watson;980114-03-9321;60178373136;S021@mail.goodbrain.edu.my;GBU2407FN
@@ -27,3 +28,7 @@ S020;Nicholas Patel;020903508054;60138591060;S020@mail.goodbrain.edu.my;GBF2407B
 =======
 S020;Nicholas Patel;020903508054;60138591060;S020@mail.goodbrain.edu.my;GBF2407BN
 >>>>>>> Stashed changes
+=======
+S020;Nicholas Patel;020903508054;60138591060;S020@mail.goodbrain.edu.my;GBF2407BNS021;Harry Watson;980114-03-9321;60178373136;S021@mail.goodbrain.edu.my;GBU2407FN
+S021;Harry;970114-07-0975;60158764297;S021@mail.goodbrain.edu.my;GBU2407FN
+>>>>>>> parent of 3af4316 (merge main into admin)
