+++ user.txt
@@ -1,4 +1,3 @@
-<<<<<<< HEAD
 <<<<<<< Updated upstream
 A001,000816a001,1
 A002,471015a002,1
@@ -42,9 +41,6 @@ S021;980114S021;3
 S021;970114S021;3
 =======
 A001;000816a001;1
-=======
-A001,000816a001;1
->>>>>>> main
 A002;471015a002;1
 L001;700601l001;4
 L002;640628l002;2
@@ -80,9 +76,5 @@ S016;040209s016;3
 S017;010624s017;3
 S018;010204s018;3
 S019;050515s019;3
-<<<<<<< HEAD
 S020;020903s020;3
 >>>>>>> Stashed changes
-=======
-S020;020903s020;3
->>>>>>> main
